Create CameraController and changed Player to have a PlayerBody and a Gun
Created a camera that follows the player around and points slightly in the direction of the cursor (similar to the camera in Enter the Gungeon).

Added screen shake effect upon firing the weapon.

Changed the player model to include a PlayerBody and a Gun that pivots around the body.

Refactored Aiming to its own script outside of PlayerMovement.